requirements.txt
flask
transformers
torch
openai-whisper
requests
beautifulsoup4
